[#2] Auth 관련 파일 분리 및 리팩토링 진행
