+++ Codive/Features/Auth/Data/SocialAuthService.swift
@@ -11,44 +11,6 @@ import KakaoSDKAuth
 import KakaoSDKCommon
 import AuthenticationServices
 
-// MARK: - Auth Result
-enum AuthResult {
-    case success(AuthUser)
-    case failure(AuthError)
-}
-
-struct AuthUser {
-    let id: String
-    let email: String?
-    let name: String?
-    let provider: AuthProvider
-}
-
-enum AuthProvider {
-    case kakao
-    case apple
-}
-
-enum AuthError: Error, LocalizedError {
-    case cancelled
-    case networkError(String)
-    case userInfoError
-    case unknown(String)
-    
-    var errorDescription: String? {
-        switch self {
-        case .cancelled:
-            return "사용자가 로그인을 취소했습니다"
-        case .networkError(let message):
-            return "네트워크 오류: \(message)"
-        case .userInfoError:
-            return "사용자 정보를 가져오는데 실패했습니다"
-        case .unknown(let message):
-            return "알 수 없는 오류: \(message)"
-        }
-    }
-}
-
 // MARK: - Social Auth Service Protocol
 protocol SocialAuthServiceProtocol {
     func kakaoLogin() async -> AuthResult
